Extract ViewPanel display logic to own component beneath BeaconSummaryPanel
